requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0  # For YOLOv8
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
pytest>=7.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 